# products/wd-image-ai/wd_image_ai/routes.py
"""The product's read API: a user's images (ADR-0023, ADR-0036), under /products/wd-image-ai/.

Every query is scoped by tenant, product and user, and every link is signed fresh, so an image can
be opened any time (the links themselves expire after an hour)."""

import re
from datetime import datetime

from fastapi import APIRouter, Depends, HTTPException, Query, Response
from pydantic import BaseModel
from wd_platform_sdk import Identity, RouteDeps

from wd_image_ai.images import ImageRecord, ImageStore, PostgresImageStore

PRODUCT_ID = "wd-image-ai"
CONTENT_TYPES = {
    "png": "image/png",
    "jpg": "image/jpeg",
    "jpeg": "image/jpeg",
    "webp": "image/webp",
}


class ImageSummary(BaseModel):
    id: str
    prompt: str
    mode: str
    width: int
    height: int
    created_at: datetime
    image_url: str
    thumb_url: str


class ImagePage(BaseModel):
    images: list[ImageSummary]
    next_before: datetime | None  # pass as `before` to get the next page; null at the end


def file_name(prompt: str, ext: str) -> str:
    """A safe, readable file name from the prompt: letters, digits and hyphens only."""
    slug = re.sub(r"[^a-z0-9]+", "-", prompt.lower()).strip("-")[:48].strip("-") or "image"
    return f"{slug}.{ext}" if ext else slug


def build_routes(deps: RouteDeps, store: ImageStore | None = None) -> APIRouter:
    """`store` is injectable for tests; production reads the shared database."""
    router = APIRouter(tags=[PRODUCT_ID])

    def images() -> ImageStore:
        return store or PostgresImageStore(deps.engine)

    async def summary(image: ImageRecord) -> ImageSummary:
        assert deps.storage is not None, "object storage is not configured"
        assert image.created_at is not None
        return ImageSummary(
            id=image.id, prompt=image.prompt, mode=image.mode, width=image.width,
            height=image.height, created_at=image.created_at,
            image_url=await deps.storage.url(image.image_key),
            thumb_url=await deps.storage.url(image.thumb_key),
        )  # fmt: skip

    @router.get("/images", response_model=ImagePage)
    async def list_images(
        limit: int = Query(20, ge=1, le=50),
        before: datetime | None = None,
        identity: Identity = Depends(deps.identity),
    ) -> ImagePage:
        rows = await images().list(identity.tenant_id, identity.user_id, limit + 1, before)
        page, more = rows[:limit], len(rows) > limit
        with deps.acting_as(identity, PRODUCT_ID):
            items = [await summary(i) for i in page]
        return ImagePage(images=items, next_before=page[-1].created_at if more and page else None)

    @router.get("/images/{image_id}", response_model=ImageSummary)
    async def get_image(image_id: str, identity: Identity = Depends(deps.identity)) -> ImageSummary:
        image = await images().get(identity.tenant_id, identity.user_id, image_id)
        if image is None:  # also what another user's image looks like: no existence leak
            raise HTTPException(404, "image not found")
        with deps.acting_as(identity, PRODUCT_ID):
            return await summary(image)

    @router.get("/images/{image_id}/download")
    async def download(image_id: str, identity: Identity = Depends(deps.identity)) -> Response:
        """The image as a file the browser saves (browsers ignore `download` on links to the
        storage host, ADR-0034)."""
        assert deps.storage is not None, "object storage is not configured"
        image = await images().get(identity.tenant_id, identity.user_id, image_id)
        if image is None:
            raise HTTPException(404, "image not found")
        try:
            with deps.acting_as(identity, PRODUCT_ID):
                data = await deps.storage.get(image.image_key)
        except FileNotFoundError:
            raise HTTPException(404, "image not found") from None
        ext = image.image_key.rsplit(".", 1)[-1].lower()
        return Response(
            data,
            media_type=CONTENT_TYPES.get(ext, "application/octet-stream"),
            headers={
                "Content-Disposition": f'attachment; filename="{file_name(image.prompt, ext)}"',
                "Cache-Control": "private, no-store",
                "X-Content-Type-Options": "nosniff",
            },
        )

    @router.delete("/images/{image_id}", status_code=204)
    async def delete_image(image_id: str, identity: Identity = Depends(deps.identity)) -> Response:
        """Delete an image: its files first, then the record."""
        assert deps.storage is not None, "object storage is not configured"
        image = await images().get(identity.tenant_id, identity.user_id, image_id)
        if image is None:
            raise HTTPException(404, "image not found")
        with deps.acting_as(identity, PRODUCT_ID):
            for key in (image.image_key, image.thumb_key):
                try:
                    await deps.storage.delete(key)
                except FileNotFoundError:
                    pass
        await images().delete(identity.tenant_id, identity.user_id, image_id)
        return Response(status_code=204)

    return router